Fix the usage email as an identifier for the authentication flow and fix tests
