+++ src/main/kotlin/org/testing_survey_creator/repository/UserRepository.kt
@@ -7,5 +7,5 @@ import java.util.Optional
 
 @Repository
 interface UserRepository : JpaRepository<User, Long> {
-    fun findByUsername(username: String): Optional<User>
+    fun findByEmail(email: String): Optional<User>
 }
